setting carousel settings
